AWS company migration
In this PowerPoint, I will explain the AWS suite offerings and how they can be used to migrate a company using external storage to the cloud. This will touch on the basics of the AWS suite and the security portion of the entire migration from a cloud solutions architect's perspective.